+++ src/Configurations/EditorLayer/Editor.test.js
@@ -37,7 +37,7 @@ const renderRemoteStorageForm = ({
   onClose = jest.fn(),
   pristine = true,
   submitting = false,
-} = {}) => (render(
+} = {}) => render(
   <QueryClientProvider client={new QueryClient()}>
     <MemoryRouter>
       <Editor
@@ -50,7 +50,7 @@ const renderRemoteStorageForm = ({
       />
     </MemoryRouter>
   </QueryClientProvider>,
-));
+);
 
 describe('Editor', () => {
   it('shows Dematic SD specific fields if Dematic SD chosen', () => {
@@ -110,44 +110,6 @@ describe('Editor', () => {
       expect(query.credProperties).not.toBeInTheDocument();
       expect(query.accessionWorkflowSection).not.toBeInTheDocument();
       expect(query.returningWorkflowSection).not.toBeInTheDocument();
-    });
-  });
-
-  describe('Data synchronization schedule delay', () => {
-    const validValues = ['1', '2', String(Number.MAX_SAFE_INTEGER)];
-    const invalidValues = ['0', '-1', '3.14', '.2', '-0.2', 'abc'];
-
-    it.each(validValues)('shows no error when valid', value => {
-      renderRemoteStorageForm();
-
-      const providers = screen.getByRole('combobox', { name: /providerName/ });
-      const dematicSdOption = within(providers).getByRole('option', { name: DEMATIC_SD });
-      user.selectOptions(providers, dematicSdOption);
-
-      const region = screen.getByRole('region', { name: /synchronization/ });
-      const input = within(region).getByRole('spinbutton');
-
-      user.type(input, value);
-      user.tab();
-
-      expect(region).not.toHaveTextContent(/synchronization.schedule.info.notValid/);
-    });
-
-    it.each(invalidValues)('shows error message when not valid', value => {
-      renderRemoteStorageForm();
-
-      const providers = screen.getByRole('combobox', { name: /providerName/ });
-      const dematicSdOption = within(providers).getByRole('option', { name: DEMATIC_SD });
-      user.selectOptions(providers, dematicSdOption);
-
-      const region = screen.getByRole('region', { name: /synchronization/ });
-      const input = within(region).getByRole('spinbutton');
-
-      user.type(input, value);
-      user.tab();
-
-      const alert = within(region).getAllByRole('alert')[0];
-      expect(alert).toHaveTextContent(/synchronization.schedule.info.notValid/);
     });
   });
 });
